requirements.txt
Flask>=2.0.0
tensorflow>=2.10.0
Pillow
numpy
opencv-python
werkzeug
gunicorn